--- log_analyzer/src/channel_calc.py
@@ -1,5 +1,0 @@
-
-
-class ChannelCalc(object):
-    def __init__(self, channel):
-        self.__channel = channel
