+++ log_analyzer/src/checkers/drop_rate_checker.py
@@ -1,5 +1,5 @@
 from src.checkers.base_checker import BaseChecker
-from src.channel_manager import ChannelManager
+from src.controllers.channels_manager import ChannelsManager
 from src.constants import DROP_RATE_THRESHOLD, HIGH_DROP_RATE
 
 
@@ -20,7 +20,7 @@ class DropRateChecker(BaseChecker):
             none
        """
     def run(self):
-        for channel in ChannelManager().get_channels():
+        for channel in ChannelsManager().get_channels():
             drop_rate = channel.get_measure(self._measure_key)
             if(drop_rate > self._parameters[DROP_RATE_THRESHOLD]):
                 channel.update_flag(HIGH_DROP_RATE, True)
