+++ log_analyzer/src/checkers/drop_ratio_checker.py
@@ -1,5 +1,5 @@
 from src.checkers.base_checker import BaseChecker
-from src.channel_manager import ChannelManager
+from src.controllers.channels_manager import ChannelsManager
 from src.constants import DROP_RATIO_THRESHOLD, HIGH_DROP_RATIO, \
     PACKET_DROPPED, READ_TIME
 
@@ -20,7 +20,7 @@ class DropRatioChecker(BaseChecker):
         none
    """
     def run(self):
-        for channel in ChannelManager().get_channels():
+        for channel in ChannelsManager().get_channels():
             drop_ratio = channel.get_measure(self._measure_key)
             if(drop_ratio > self._parameters[DROP_RATIO_THRESHOLD]):
                 channel.update_flag(HIGH_DROP_RATIO, True)
