+++ log_analyzer/src/checkers/execution_time_checker.py
@@ -1,8 +1,7 @@
 from src.checkers.base_checker import BaseChecker
-from src.channel_manager import ChannelManager
-from src.channel_calc import ChannelCalc
-from src.constants import (HIGH_EXECUTION_TIME, EXECUTION_TIME_THRESHOLD,
-                           EXECTION_TIME)
+from src.constants import (HIGH_EXECUTION_TIME, EXECUTION_TIME_THRESHOLD)
+from src.controllers.channels_manager import ChannelsManager
+
 
 # This class contains the logic to check execution time and to set the execution time flag of the
 # channels
@@ -21,7 +20,7 @@ class ExecutionTimeChecker(BaseChecker):
    """
 
     def run(self):
-        for channel in ChannelManager().get_channels():
+        for channel in ChannelsManager().get_channels():
             execution_time = channel.get_measure(self._measure_key)
             if (execution_time > self._parameters[EXECUTION_TIME_THRESHOLD]):
                 channel.update_flag(HIGH_EXECUTION_TIME, True)
