+++ log_analyzer/src/checkers/read_time_checker.py
@@ -1,5 +1,5 @@
 from src.checkers.base_checker import BaseChecker
-from src.channel_manager import ChannelManager
+from src.controllers.channels_manager import ChannelsManager
 from src.constants import CHANNEL_READ_THRESHOLD, HIGH_READ_TIME, READ_TIME
 
 
@@ -19,7 +19,7 @@ class ReadTimeChecker(BaseChecker):
          none
     """
     def run(self):
-        for channel in ChannelManager().get_channels():
+        for channel in ChannelsManager().get_channels():
             read_time = channel.get_measure(self._measure_key)
             if (read_time > self._parameters[CHANNEL_READ_THRESHOLD]):
                 channel.update_flag(HIGH_READ_TIME, True)
