+++ log_analyzer/src/checkers/time_to_buffer_average_checker.py
@@ -1,7 +1,6 @@
 from src.checkers.base_checker import BaseChecker
-from src.channel_manager import ChannelManager
-from src.constants import CHANNEL_TIME_TO_BUFFER, HIGH_CHANNEL_TIME_TO_BUFFER,\
-    CHANNEL_TIME_TO_BUFFER_AVERAGE_THRESHOLD, PACKET_RETRIEVED, EXECTION_TIME
+from src.controllers.channels_manager import ChannelsManager
+from src.constants import HIGH_CHANNEL_TIME_TO_BUFFER, CHANNEL_TIME_TO_BUFFER_AVERAGE_THRESHOLD
 
 
 # This class contains the logic to check time to buffer of the channels
@@ -18,7 +17,7 @@ class TimeToBufferAverageChecker(BaseChecker):
        Returns: none
     """
     def run(self):
-        for channel in ChannelManager().get_channels():
+        for channel in ChannelsManager().get_channels():
             time_to_buffer = channel.get_measure(self._measure_key)
             if (time_to_buffer > self._parameters[CHANNEL_TIME_TO_BUFFER_AVERAGE_THRESHOLD]):
                 channel.update_flag(HIGH_CHANNEL_TIME_TO_BUFFER, True)
