+++ log_analyzer/src/controllers/checkers_manager.py
